Improved use of localStorage to navegate through Registro with Atras btn #259

+++ frontend/src/views/asesorado/registroAsesoradoResumen/RegistroAsesoradoResumen.jsx
@@ -2,9 +2,9 @@ import React, { useEffect } from 'react'
 
 import { TemplateRegistroUsuario } from '../../../routeIndex'
 
-import data from './data.json';
-
 import { useNavigate } from "react-router-dom";
+
+import noUserImg from '../../../assets/noUserImg.png'
 
 let progressBar = {
     "currentStep": 2,
@@ -34,38 +34,34 @@ const RegistroAsesoradoResumen = () => {
 
     let navigate = useNavigate()
 
+    const nombreUser = localStorage.registro1_nombre
+    const apellidoPatUser = localStorage.registro1_apellidoPaterno
+    const apellidoMatUser = localStorage.registro1_apellidoMaterno
+    const matriculaUser = localStorage.registro1_matricula
+    const carreraUser = localStorage.registro1_carrera
+    const telefonoUser = localStorage.registro1_telefono
+    const imagenUser = localStorage.registro1_imagenPerfil
+
     // Verificamos que se cuente con todos campos obligatorios
     // Si no, se envía al usuario a la landing page
     useEffect(() => {
-        if(!localStorage.registro1_nombre || 
-        !localStorage.registro1_apellidoPaterno || 
-        !localStorage.registro1_matricula || 
-        !localStorage.registro1_carrera) {
+        if(!nombreUser || !apellidoPatUser || !matriculaUser || !carreraUser) {
             localStorage.clear()
             navigate('/landingPage')
         }
-    }, [navigate])
-    
-
-    //función para armar cadena del campo cadenas, retorna un String armado
-    const armarCadenaCarrea = (carreras) => {
-        let cadena ="";
-        for( let i = 0; i<carreras.length ; i++ ){
-            if( i === carreras.length - 1 )
-                cadena += carreras[i];
-            else
-            cadena += carreras[i] + ', ';
-
-        }
-        return cadena;
-    }
-
+    })
 
     return (
     
         <TemplateRegistroUsuario 
             progressBarJSON={progressBar}
             btnAtrasRoute ="./registroAsesoradoCondiciones"
+            btnSiguienteProps={ 
+                {
+                    view: 3, 
+                    props: { nombreUser, apellidoPatUser, apellidoMatUser, matriculaUser, carreraUser, telefonoUser, imagenUser }
+                } 
+            }
             ultimoTexto='Confirmar'>
             
             <div className='bloq_condicionesAsesorado'>
@@ -79,8 +75,11 @@ const RegistroAsesoradoResumen = () => {
 
                         <div className='contenedor-img-perfil-InfPerUsuario'>
 
-                            <img src={ require('../../../assets/img-editar-perfil-tmp.JPG')  } alt="imgProfile" className='imagen-InfPerUsuario'/>
-                            
+                            <img src={ imagenUser ? imagenUser : noUserImg } 
+                                alt="imgProfile" 
+                                className='imagen-InfPerUsuario'
+                                style={{objectFit: 'cover'}}
+                            />
 
                         </div>
 
@@ -88,30 +87,20 @@ const RegistroAsesoradoResumen = () => {
 
                     <div className='contenedor-InfPerUsuario-der' >
                         
-                        <p className='etiqueta-nombre-InfPerUsuario' >{ data["informacion"].camposObligatorios[0].info }</p>
-                        <p className='etiqueta-matricula-InfPerUsuario' >{ data["informacion"].camposObligatorios[1].info }</p>
+                        <p className='etiqueta-nombre-InfPerUsuario' >
+                            { nombreUser + " " + apellidoPatUser }
+                        </p>
+                        <p className='etiqueta-matricula-InfPerUsuario' >{ matriculaUser }</p>
 
-                        {
-                        // iteramos con map entre los camposVariantes que tenemos en el archivo json
-                        data["informacion"].camposVariantes.map((campos) => (
-                            
-                            <p className={ `etiqueta-${campos.nombreClase}-InfPerUsuario` }  >
-
-                                { 
-                                    //condicional para arreglar cadena de carreras
-                                    ( campos.campo === "Carrera(s)" ?
-                                        (campos.info.length === 1 ? 
-                                            `${ campos.campo }: ${ campos.info[0]}`
-                                            : 
-                                            `${campos.campo}: ${ armarCadenaCarrea(campos.info) }`  )
-                                        : 
-                                        `${campos.campo}: ${campos.info}`
-                                    )  
-                                }
-
-                            </p>
-                        )) 
-                        }
+                        <p className={ 'etiqueta-carrera-InfPerUsuario' }  >
+                            <span style={{fontWeight: '600'}}>Carrera: </span> 
+                            { carreraUser ? carreraUser.slice(0, carreraUser.indexOf(" ")) : null }
+                        </p>
+                        
+                        <p className={ 'etiqueta-telefono-InfPerUsuario' }  >
+                            <span style={{fontWeight: '600'}}>Teléfono: </span> 
+                            { telefonoUser ? telefonoUser : 'No se ingresó' }
+                        </p>
 
                     </div>
